chore: update getTSQualifiedNameFromClientName to handle DynamoDB DocumentClient
